intraday/requirements.txt
google-cloud-storage==1.35.0